# backend/openapi.yaml
openapi: 3.0.3
info:
  title: EarthSync Server API
  description: API for managing users, key exchange, data ingest, and historical data retrieval (spectrograms and peaks). Includes enhanced server-side peak detection and basic transient detection. Archives older data to database. Returns structured, timestamped spectrogram history.
  version: 1.1.14 # Version bump reflects history API change
servers:
  - url: http://localhost:3000 # Modified to remove backticks for compatibility
    description: Development Server (HTTP)
paths:
...
  /key-exchange:
    post:
      summary: Exchange encryption key for WebSocket communication
      tags: [WebSocket]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/KeyExchangeSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /users/{username}:
    delete:
      summary: Delete a user account (self-deletion)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: username
          in: path
          required: true
          # Corrected pattern: Removed unnecessary backslash escape for $
          schema: { type: string, minLength: 3, maxLength: 30, pattern: '^[a-zA-Z0-9_]+$' }
          description: The username of the account to delete. Must match the authenticated user.
      responses:
        '200': { $ref: '#/components/responses/UserDeleteSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          {
            type: string,
            description: 'A description of the error.',
            example: 'Invalid username or password',
          }
      required: [error]
  responses:
...
    KeyExchangeSuccess:
      description: Key exchanged successfully. Returns the encryption key.
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties:
                    {
                      key:
                        {
                          type: string,
                          description: 'AES-256 encryption key (hex, 64 chars)',
                          example: 'a1b2c3d4e5f6...',
                        },
                    },
                },
            },
        }
...
    UserDeleteSuccess:
      description: User deleted successfully.
      content:
        {
          application/json:
            {
              schema:
                {
                  type: object,
                  properties: { message: { type: string, example: User deleted successfully } },
                },
            },
        }
...
    BadRequest:
      description: Invalid input provided (e.g., failed validation, incorrect format).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unauthorized:
      description: Authentication failed (e.g., missing/invalid/expired JWT token).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Forbidden:
      description: Authorization failed (e.g., invalid API key, attempting action without permission).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: The requested resource could not be found.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
    TooManyRequests:
      description: Rate limit exceeded for the requested endpoint.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    InternalServerError:
      description: An unexpected error occurred on the server.
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from the /login endpoint. Include in the Authorization header as 'Bearer {token}'."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'API Key required only for the /data-ingest endpoint.'
...